Fix dotfiles

+++ Makefile
@@ -11,17 +11,17 @@ packages:
 
 dotfiles:
 	# add aliases for dotfiles
-	for file in $(shell find $(CURDIR) -name ".*" -not -name ".gitmodules" -not -name ".gitignore" -not -name ".git" -not -name ".*.swp"); do \
+	for file in $(shell find $(CURDIR) -name ".*" -not -name ".git" -not -name ".*.swp" -not -name ".travis.yml"); do \
 		f=$$(basename $$file); \
 		ln -sfn $$file $(HOME)/$$f; \
 	done
 	# Non public dotfiles
-	for file in $(shell find ~/Dropbox/DOTFILES -name ".*" -not -name ".gitmodules" -not -name ".gitignore" -not -name ".git" -not -name ".*.swp"); do \
+	for file in $(shell find ~/Dropbox/DOTFILES -maxdepth 1 -name ".*"); do \
 		f=$$(basename $$file); \
 		ln -sfn $$file $(HOME)/$$f; \
 	done
 
-	ln -s $(CURDIR) ~/.dotfiles
+	if [ -f ~/.dotfiles ] ; then ln -s $(CURDIR) ~/.dotfiles; fi
 
 sshkey:
 	ssh-add -K ~/.ssh/id_rsa
